Quic API and implementation.
Motivation:

Quic can be properly supported since it is out of incubation from Netty, it provides a new transport for building middlewares on top of it. Although HTTP/3 is currently the only protocol specified on top of Quic, there are discussions and will to port other protocols on top of Quic.

While the purpose is similar to NetClient/NetServer there are a few significant differences that would make the `io.vertx.core.net` API reuse.

- QUIC mandates TLS and does not support TLS upgrade
- QUIC requires to bind a _physical_ UDP socket on client and server prior other interactions
- QUIC provides connections multiplexing streams
- A connection can be closed with a reason
- Streams can be single duplex or full duplex
- Server load balancing operates in a different fashion

Vert.x can facilitate the usage of Quic in the Java ecosystem by providing a comprehensive solution and explore ways of using Netty codecs on top of Quic.

This feature provides an API to interact with Quic transport as a server or a client, very much like the existing TCP socket API. It also provides an internal API to interact with the underlying Netty configuration to plugin codecs and use messages instead of buffers.

Changes:

Quic API and implementation.