Get rid of VertxSslContext
Motivation:

Our implementation has maintained a VertxSslContext wrapping an SslContext to configure a few aspects of the SSL engine. This does not seem necessary anymore since SslContext has all the necessary configuration options on the SslContextBuilder.

Changes:

Update the SslContextFactory to handle the configuration aspects VertxSslContext was implementing.

Remove VertxSslContext wrapper and configure the options instead on the SslContextFactory.