Let the HttpChannelConnector become an interface and embbed only the resource it needs to.
Motivation:

The HttpClient implementations relies on the HttpChannelConnector to obtain a configured HTTP connection.

While the HttpChannelConnector does not leak to HttpClient, the creation and usage of HttpChannelConnector makes it hard to swap with another implementations (let's say an HTTP/3 version).

HttpChannelConnector references resources (NetClient transport / HttpClientMetrics) as well as configuration (various options), the presence of options implies to create connector instances per option. Ideally we should have a single connector per client and pass options as arguments when a connection is needed (which is fine because the connector does not actually cache state).

Changes:

Let HttpChannelConnector become an interface defining its needs.

The HTTP/1.x / H2 implementations becomes an implementation and references the TCP client it requires.

Usage of HttpClientImpl to obtain configuration is removed, so the connector can be created and passed to HttpClientImpl (in other words, we should avoid something like new HttpConnector(this.something)).