Quic stream reset handling improvements.
Motivation:

The initial implementation of stream reset handling closes the stream upon reception of a reset frame when a stream handler is present.

It is the desired default behavior, however a stream having sent or received a reset frame can still process or read data on the other direction. Setting a handler on the stream makes the application aware of the reset and when it happens the stream should not be closed (that is writing a final stream frame) and instead the application should take care of it at the most appropriate time the protocol requires it.

Changes:

Setting a quic stream handler inhibates exception handling and stream closing.