Introduce HttpCompressionOptions.
Motivation:

HTTP/3 will require configure of compression. This introduces a new HttpCompressionOptions configuration that encapsulate anything related to compression configuration.

Changes:

The HttpServerConfiguration is modified to use this setting and delegate the current related methods to the underlying compression object to support the transition.

HTTP compression tests have been reworked to use the new HTTP configuration mechanism and reduce the proliferation of test subclasses to reduce the amount of necessary code to test the variety of algorithms.