NetSocket.exceptionHandler reports UnsupportedOperationException on close (#5823)
Follows-up on https://github.com/eclipse-vertx/vert.x/pull/5783

If the user closes the connection and then the client is closed, UnsupportedOperationException is reported to the socket exception handler.

Instead of throwing UOE, the redundant shutdown event coming is ignored.

Also, make sure to complete the channel promise when the channel is closed.
Otherwise, the connection close future won't be completed.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>