Provide an extension of the HttpClientRequest API that exposes the request metric.
Motivation:

Sometimes it is necessary to let framework access to underlying metric of an HTTP request in order to reliably and efficiently share state with the metrics implementation.

This is useful for frameworks that build on top of HTTP client and wants to extend metrics reporting beyond Vert.x capabitlies.