Extend ClientMetrics SPI to introduce a new init method that is called by the HTTP client before the request is actually sent.
Motivation:

The ClientMetrics SPI design creates a request metric when the request headers are written, we should actually relax this constraint and let client obtain a metric from the SPI before headers are written, in order to give the oppportunity to customize the request metric object before headers are written.

Changes:

Extend ClientMetrics SPI to add a new init method that produces a metrics instance.

When the implementation returns null, the client behaves with a compatibility mode and calls ClientMetrics#requestBegin to signal the beginning of the request and obtain the request metric.

When the implementation returns non-null, the client instead calls the ClientMetrics#requestBegin overload that accepts the request metric instance that should be handed back to the SPI.