Add HTTP client request protocol selection.
Motivation:

The HTTP client request protocol is selected by the HttpClientOptions. With the advent of alternative services we should support specifying the protocol per request instead since a server might indicate an alternative server for a given protocol.

Changes:

HttpConnectOptions is added a protocol property that defines the desired protocol for the request, which is by default null, when null the client default protocol is used.

HTTP protocol is added to the client pool key so we distinguish a connection by its desired version, allowing the client to maintain connections to a given server for different versions of the protocol.